Housekeeping 1 - Basic template
Each page now has consistent format:

The header is the same across all pages and the footer correctly nests at the bottom of each page. 

This was achieved by adding a placeholder "main" section and pushing it to the bottom by padding at the top. 

Can now use this to create a test branch and start to experiment with adding features